requirements.txt
fastapi
uvicorn
sqlalchemy
streamlit
python-dotenv
imagekitio[aiohttp]
aiosqlite
fastapi-offline
python-multipart
fastapi_users
fastapi-users-db-sqlalchemy